requirements.txt
# package
flask_restx==1.3.0 
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0 
SQLAlchemy==2.0.25 
sqlalchemy-utils==0.41.1
python-dotenv==1.0.1
alembic==1.13.1
psycopg2-binary==2.9.9
boto3==1.34.32 
python-jose==3.3.0 


# customize 